refactor(components): 分离AI聊天、连接和新闻组件
